Move text input to widget module
